+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [Add support for geometry fields in CSV exports #884](https://github.com/farmOS/farmOS/pull/884)
+
 ### Fixed
 
 - [Fix config export via UI #883](https://github.com/farmOS/farmOS/pull/883)